% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nycdogs}
\alias{nycdogs}
\title{NYC Dog Licensing Data set}
\format{
\subsection{\code{nycdogs}}{

A data frame with 508,196 rows and 8 columns:
\describe{
\item{animal_name}{User-provided dog name (unless specified otherwise)}
\item{animal_gender}{M (Male) or F (Female) dog gender}
\item{animal_birth_year}{Year dog was born}
\item{breed_name}{Dog breed}
\item{zip_code}{Owner zip code}
\item{license_issued_date}{Date the dog license was issued}
\item{license_expired_date}{Date the dog license expires}
\item{extract_year}{Year the data was extracted}
}
}
}
\source{
\url{https://data.cityofnewyork.us/Health/NYC-Dog-Licensing-Dataset/nu7n-tubp}
}
\usage{
nycdogs
}
\description{
NYC dog licensing data from the July 1, 2022 update.
}
\keyword{datasets}
